service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.25.0
pillow>=10.0.0
numpy>=1.24.0
faiss-cpu>=1.11.0
open-clip-torch>=2.20.0
requests>=2.25.0
pydantic>=2.0.0
google-cloud-storage>=2.10.0
